Keep tmux overlay flash helpers under budget
